en views agregue decoradores y un condicional para editar y eliminar, en home.html agreaue un condicional para mostrar botones editar y eliminar

+++ base/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, redirect
 from django.http  import HttpResponse
 from .models import Room, Topic
 from .forms import RoomForm
+from  django.contrib.auth.decorators import login_required
 from django.db.models  import Q
 from django.contrib.auth.models import User
 from django.contrib import messages
@@ -74,6 +75,7 @@ def room(request,pk):
     return  render (request,'base/room.html', context)
 
 
+@login_required(login_url='login')
 def createRoom(request):
      form=RoomForm()
      if request.method=='POST':
@@ -86,10 +88,14 @@ def createRoom(request):
      context={'form':form}
      return  render (request,'base/room_form.html', context)
 
-    
+
+@login_required(login_url='login')   
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return  HttpResponse("Your are not allower here!")
 
     if request.method == 'POST':
        form = RoomForm(request.POST, instance=room)
@@ -100,8 +106,14 @@ def updateRoom(request, pk):
     context = {'form': form}   
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='login')   
 def deleteRoom(request,pk):
     room=Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return  HttpResponse("Your are not allower here!")
+ 
     if request.method == 'POST':
         room.delete()
         return redirect('home')
